Initial commit of inb4404.py (in development)
This is an enhanced version of 'download.py'. With this version it is possible to give a filename instead of a thread-url. In this file you save all the thread urls you want to download (one url per line). Lines starting with '#' are ignored.

Please keep in mind that this is still in development.