fixed file reloading, reloads every 5 minutes now
Not quite sure why I added the block from line 96 to line 101, but it seems like that it prevents file reloading as it is intended. In addition to that, I changed the feature so that it reloads every 5 minutes. In case it's still not working as intended, please report the issue.